# backend/src/swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "Space App",
      "description": "Final fullstack project for Okten School",
      "version": "1.0.0"
   },
   "servers": [
      {
         "url": "http://localhost:5400/api",
         "description": "Express"
      }
   ],
   "paths": {
...
      "/plans": {
         "get": {
            "summary": "get all plans",
            "tags": [
               "Plans"
            ],
            "operationId": "getAllPlans",
            "parameters": [
               {
                  "in": "query",
                  "name": "searchKey",
                  "schema": {
                     "type": "string"
                  },
                  "description": "Keyword for searching"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Plans"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/plans/add": {
         "get": {
            "summary": "add plan",
            "tags": [
               "Plans"
            ],
            "operationId": "addPlan",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "201": {
                  "description": "New plan was created",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Plan"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/plans/{planId}": {
         "get": {
            "summary": "get plan by id",
            "tags": [
               "Plans"
            ],
            "operationId": "getPlanById",
            "parameters": [
               {
                  "name": "planId",
                  "in": "path",
                  "required": true,
                  "description": "plan id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Plan"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "summary": "update plan by id",
            "tags": [
               "Plans"
            ],
            "operationId": "updatePlanById",
            "parameters": [
               {
                  "name": "planId",
                  "in": "path",
                  "required": true,
                  "description": "plan id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "title": {
                              "type": "string",
                              "example": "English Plan"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "summary": "delete plan by id",
            "tags": [
               "Plans"
            ],
            "operationId": "deletePlanById",
            "parameters": [
               {
                  "name": "planId",
                  "in": "path",
                  "required": true,
                  "description": "plan id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "securitySchemes": {
         "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "schemas": {
         "UnauthorizedError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 401
               },
               "message": {
                  "type": "string",
                  "example": "Invalid or expired token"
               }
            }
         },
         "ObjectIdError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 400
               },
               "message": {
                  "type": "string",
                  "example": "Object ID is not valid"
               }
            }
         },
         "NotExistError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 404
               },
               "message": {
                  "type": "string",
                  "example": "Object does not exist"
               }
            }
         },
         "DefaultError": {
            "type": "object",
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 500
               },
               "message": {
                  "type": "string",
                  "example": "Error message"
               }
            }
         },
         "Success": {
            "type": "object",
            "properties": {
               "message": {
                  "type": "string",
                  "example": "Success"
               }
            }
         },
...
         "Plan": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "string",
                  "example": "63dfe16eda233c96fc6e2604"
               },
               "title": {
                  "type": "string",
                  "example": "Training"
               },
               "lastModified": {
                  "type": "number",
                  "example": 1677765747660
               }
            }
         },
         "Plans": {
            "type": "array",
            "items": {
               "$ref": "#/components/schemas/Plan"
            }
         },
...
      }
   }
}